+++ src/lib/sqlite.ts
@@ -4,7 +4,6 @@ export type Database = NativeSqliteDatabase;
 
 type DatabaseOptions = {
 	readonly?: boolean;
-	fileMustExist?: boolean;
 };
 
 type PragmaOptions = {